# openApi.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "0.0.1",
    "title": "Signal K Seamap Plugin API",
    "description": "Global sea charts based on OSM/Gebco/Emod with contour generation and bathymetry support",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers": [
    {
      "url": "/plugins/signalk-seamap-plugin",
      "description": "Plugin base path"
    }
  ],
  "paths": {
...
    "/bathymetry/{name}.json": {
      "get": {
        "tags": ["Bathymetry"],
        "summary": "Get bathymetry TileJSON metadata",
        "description": "Returns TileJSON metadata for bathymetry depth contours",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "DEM source name (mapterhorn, gebco, emod)",
            "schema": {
              "type": "string",
              "enum": ["mapterhorn", "gebco", "emod"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TileJSON metadata with vector layer info",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tilejson": { "type": "string" },
                    "name": { "type": "string" },
                    "format": { "type": "string" },
                    "vector_layers": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "string" },
                          "fields": {
                            "type": "object",
                            "properties": {
                              "elevation": { "type": "string" },
                              "depth": { "type": "string" },
                              "level": { "type": "string" }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Source not found"
          }
        }
      }
    },
    "/bathymetry/{name}/{z}/{x}/{y}.pbf": {
      "get": {
        "tags": ["Bathymetry"],
        "summary": "Get bathymetry depth contours tile",
        "description": "Generates and delivers bathymetry depth contours as vector tile (MVT/PBF) with configurable depth levels",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "description": "DEM source name",
            "schema": {
              "type": "string",
              "enum": ["mapterhorn", "gebco", "emod"]
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "description": "Zoom level",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 14
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "description": "Tile X coordinate",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "description": "Tile Y coordinate",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Bathymetry vector tile",
            "content": {
              "application/x-protobuf": {}
            }
          },
          "204": {
            "description": "No tile available"
          },
          "400": {
            "description": "Invalid tile coordinates"
          }
        }
      }
    },
...
    "/composite/{provider}.json": {
      "get": {
        "tags": ["Composite"],
        "summary": "Get composite TileJSON metadata",
        "description": "Returns TileJSON metadata for composite tiles combining OSM, Seamap, Contours, Bathymetry and Soundings",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "DEM provider name for bathymetry layers (gebco, emod, mapterhorn)",
            "schema": {
              "type": "string",
              "enum": ["gebco", "emod", "mapterhorn"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "TileJSON metadata for composite source",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "tilejson": { "type": "string" },
                    "name": { "type": "string" },
                    "description": { "type": "string" },
                    "format": { "type": "string" }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Provider not found"
          }
        }
      }
    },
    "/composite/{provider}/{z}/{x}/{y}.pbf": {
      "get": {
        "tags": ["Composite"],
        "summary": "Get composite tile",
        "description": "Delivers a composite vector tile combining OSM, Seamap, Contours, Bathymetry and Soundings into a single PBF",
        "parameters": [
          {
            "name": "provider",
            "in": "path",
            "required": true,
            "description": "DEM provider name",
            "schema": {
              "type": "string",
              "enum": ["gebco", "emod", "mapterhorn"]
            }
          },
          {
            "name": "z",
            "in": "path",
            "required": true,
            "description": "Zoom level",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 14
            }
          },
          {
            "name": "x",
            "in": "path",
            "required": true,
            "description": "Tile X coordinate",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "y",
            "in": "path",
            "required": true,
            "description": "Tile Y coordinate",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Composite vector tile",
            "content": {
              "application/x-protobuf": {}
            }
          },
          "204": {
            "description": "No tile available"
          },
          "400": {
            "description": "Invalid tile coordinates"
          },
          "404": {
            "description": "Provider not found"
          }
        }
      }
    }
  },
...
}